refactor: seperate ImageUploader.tsx
